requirements.txt
msal
requests
azure-identity
apache-beam
azure-storage-file-datalake
pandas
